fix: should respect visibility timeout of each consumer

+++ core/task_mark_running_as_retryable_or_discarded.sql
@@ -5,7 +5,7 @@ SET
     WHEN attempt_count >= @attempt_max THEN @discarded_state::SMALLINT
     ELSE @retryable_state::SMALLINT END,
   finalized_at = CASE
-    WHEN attempt_count >= @attempt_max THEN @finalized_at::BIGINT
+    WHEN attempt_count >= @attempt_max THEN @finalized_at
     ELSE finalized_at END
 WHERE 
   event_id IN (%s) AND state = @running_state::SMALLINT
